% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load CEFI data}
\usage{
load_data(inurl)
}
\arguments{
\item{inurl}{Thredds data url}
}
\value{
Time sliced data structure
}
\description{
This function lazy loads CEFI data from a thredds url and reads a time slice into memory.
}
